Moving the System.Globalization contract into System.Runtime
