+++ src/System.Collections/tests/project.json
@@ -2,7 +2,6 @@
   "dependencies": {
     "Microsoft.NETCore.Platforms": "1.0.2-beta-24327-03",
     "System.Collections": "4.0.12-beta-24327-03",
-    "System.Globalization": "4.0.12-beta-24327-03",
     "System.IO": "4.1.1-beta-24327-03",
     "System.Linq": "4.1.1-beta-24327-03",
     "System.Linq.Expressions": "4.1.1-beta-24327-03",
