+++ src/System.Runtime/tests/Performance/System.Runtime.Performance.Tests.csproj
@@ -37,6 +37,8 @@
     <None Include="project.json" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\..\..\System.Globalization\pkg\System.Globalization.pkgproj">
+    </ProjectReference>
     <ProjectReference Include="..\..\pkg\System.Runtime.pkgproj">
       <Project>{1e689c1b-690c-4799-bde9-6e7990585894}</Project>
       <Name>System.Runtime</Name>
