+++ src/System.Runtime/tests/Performance/project.json
@@ -3,7 +3,6 @@
     "Microsoft.NETCore.Platforms": "1.0.2-beta-24327-03",
     "System.Collections.NonGeneric": "4.0.2-beta-24327-03",
     "System.Console": "4.0.1-beta-24327-03",
-    "System.Globalization": "4.0.12-beta-24327-03",
     "System.Linq.Expressions": "4.1.1-beta-24327-03",
     "System.ObjectModel": "4.0.13-beta-24327-03",
     "System.Reflection": "4.1.1-beta-24327-03",
